test/icaresystem/Model/AppointmentTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package icaresystem.Model;

import org.junit.AfterClass;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author Joseph Hackett
 */
public class AppointmentTest {
    
    public AppointmentTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
    }
    
    @AfterClass
    public static void tearDownClass() {
    }

    /**
     * Test of requestCheckup method, of class Appointment.
     */
    @Test
    public void testRequestCheckup() {
        System.out.println("requestCheckup");
        Appointment instance = new Appointment();
        instance.requestCheckup();
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of getAppointment method, of class Appointment.
     */
    @Test
    public void testGetAppointment() {
        System.out.println("getAppointment");
        Appointment instance = new Appointment();
        int expResult = 0;
        int result = instance.getAppointment();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of setAppointment method, of class Appointment.
     */
    @Test
    public void testSetAppointment() {
        System.out.println("setAppointment");
        int appointment = 0;
        Appointment instance = new Appointment();
        instance.setAppointment(appointment);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of getPatient method, of class Appointment.
     */
    @Test
    public void testGetPatient() {
        System.out.println("getPatient");
        Appointment instance = new Appointment();
        Patient expResult = null;
        Patient result = instance.getPatient();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of setPatient method, of class Appointment.
     */
    @Test
    public void testSetPatient() {
        System.out.println("setPatient");
        Patient patient = null;
        Appointment instance = new Appointment();
        instance.setPatient(patient);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of getAppointmentDescription method, of class Appointment.
     */
    @Test
    public void testGetAppointmentDescription() {
        System.out.println("getAppointmentDescription");
        Appointment instance = new Appointment();
        String expResult = "";
        String result = instance.getAppointmentDescription();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of setAppointmentDescription method, of class Appointment.
     */
    @Test
    public void testSetAppointmentDescription() {
        System.out.println("setAppointmentDescription");
        String appointmentDescription = "";
        Appointment instance = new Appointment();
        instance.setAppointmentDescription(appointmentDescription);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }
    
}
